requirements.txt
openai
pinecone
pandas
streamlit
python-dotenv
uuid
numpy